f16 decompression in tinygrad
